+++ CompetitionTemplate/src/competition/Robot.java
@@ -1,17 +1,10 @@
 
 package competition;
 
-import competition.operatorinterface.OperatorInterfaceCommandMap;
+import competition.operator_interface.OperatorInterfaceCommandMap;
 import competition.subsystems.SubsystemDefaultCommandMap;
 import xbot.common.wpi_extensions.BaseRobot;
 
-/**
- * The VM is configured to automatically run this class, and to call the
- * functions corresponding to each mode, as described in the IterativeRobot
- * documentation. If you change the name of this class or the package after
- * creating this project, you must also update the manifest file in the resource
- * directory.
- */
 public class Robot extends BaseRobot {
 
 	@Override
