+++ CompetitionTemplate/src/competition/operator_interface/OperatorInterface.java
@@ -1,4 +1,4 @@
-package competition.operatorinterface;
+package competition.operator_interface;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
@@ -11,12 +11,14 @@ import xbot.common.wpi_extensions.mechanism_wrappers.XJoystick;
  * interface to the commands and command groups that allow control of the robot.
  */
 @Singleton
-public class OI {
+public class OperatorInterface {
 	public XJoystick leftJoystick;
 	public XJoystick rightJoystick;
 	
+	public JoystickButtonManager leftButtons; 
+	
 	@Inject
-	public OI(WPIFactory factory) {
+	public OperatorInterface(WPIFactory factory) {
 		leftJoystick = factory.getJoystick(1);
 		rightJoystick = factory.getJoystick(2);
