+++ CompetitionTemplate/src/competition/subsystems/drive/commands/TankDriveWithJoysticksCommand.java
@@ -2,17 +2,17 @@ package competition.subsystems.drive.commands;
 
 import com.google.inject.Inject;
 
-import competition.operatorinterface.OI;
+import competition.operator_interface.OperatorInterface;
 import competition.subsystems.drive.DriveSubsystem;
 import xbot.common.wpi_extensions.BaseCommand;
 
 public class TankDriveWithJoysticksCommand extends BaseCommand {
 
 	final DriveSubsystem driveSubsystem;
-	final OI oi;
+	final OperatorInterface oi;
 	
 	@Inject
-    public TankDriveWithJoysticksCommand(OI oi, DriveSubsystem driveSubsystem)
+    public TankDriveWithJoysticksCommand(OperatorInterface oi, DriveSubsystem driveSubsystem)
     {
         this.oi = oi;
         this.driveSubsystem = driveSubsystem;
